src/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
aiosqlite==0.21.0
alembic==1.16.4
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asyncpg==0.30.0
attrs==25.3.0
bcrypt==4.1.2
billiard==4.2.1
celery==5.5.3
certifi==2025.7.14
cffi==1.17.1
cfgv==3.4.0
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==45.0.5
distlib==0.4.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.4
fastapi-users==14.0.1
fastapi-users-db-sqlalchemy==7.0.0
filelock==3.18.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
identify==2.6.12
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kombu==5.5.4
loguru==0.7.3
makefun==1.16.0
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.6.3
nodeenv==1.9.1
orjson==3.11.1
packaging==25.0
passlib==1.7.4
platformdirs==4.3.8
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.2
pwdlib==0.2.1
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-telegram-bot==22.3
PyYAML==6.0.2
redis==6.4.0
rich==14.1.0
rich-toolkit==0.14.8
rignore==0.6.4
rsa==4.9.1
ruff==0.11.11
sentry-sdk==2.33.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ujson==5.10.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0; sys_platform != "win32"
vine==5.1.0
virtualenv==20.32.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
loguru==0.7.3
python_telegram_bot==22.3
tzdata>=2024.1
yarl==1.20.1
flower==2.0.1
sqladmin==0.21.0